Add files via upload
DatosMH_X.txt contains the input data (the nine okada parameters for each simulation). DatosMH_Y.txt contains the corresponding values for the maximum height in each location we want to predict. Each line of both files corresponding to the same simulation, i.e., line 1 in DatosMH_X.txt is the input for the first simulation, and line 1 in DatosMH_Y.txt is the output for the first simulation.

DatosMH_Y.txt has 6 columns, each one corresponding to a location.